// movie-client/src/models/IMovies.ts
export default interface IMovies {
  id: number;
  title: string;
  year: number;
  genres: [];
  ratings: [];
  poster: string;
  contentRating: string;
  duration: string;
  releaseDate: string;
  averageRating: number;
  originalTitle: string;
  storyline: string;
  actors: [];
  imdbRating: number;
  posterurl: string;
}